\name{reportBug}
\alias{reportBug}
\title{Function to Report Bugs in MuTossGUI}
\description{
  This function opens a graphical user interface for reporting bugs with the mutoss Java GUI.
}
\usage{
  reportBug()
}
\value{
	None (invisible \code{NULL}).
}
\details{
  This function collects all necessary information for debugging from the current session.
  The information that is sent can be reviewed before sending.
}
\author{Kornelius Rohmeyer \email{rohmeyer@small-projects.de}}
\examples{
  \dontrun{
  reportBug()
  }
}
\keyword{manip}