\name{mutossGUI}
\alias{mutossGUI}
\title{Function to start the MuTossGUI Java class}
\description{
  This function executes a method of a Java class
  defined in a JAR inside the R package.
}
\usage{
  mutossGUI()
}
\details{
  Details are given in the package vignette.
}
\value{
	None (invisible \code{NULL}).
}
\author{MuToss Coding Team}
\examples{
  \dontrun{mutossGUI()}
}
\keyword{manip}